Integrate the ipc_plugin module
This commit adds the module to load plugins and include an IPC
module to receive/send signals.

It adds the run_ipc() method invoked by the --ipc command line
option in the main.py script to execute the program using the
pluggable IPC module.

It also includes the test_sender.py unittest to test the plugin
loader and sending messages using the IPC module.

Signed-off-by: Luis Araujo <luis.araujo@collabora.co.uk>