Integrate IPC pluggable infrastructure
This commit integrates the IPC pluggable infrastructure
and an IPC (ZeroMQ) module that can be loaded using such
a mechanism.

It extends the vsm script to accept the '--ipc-module'
option to switch between the default stdin/stodut module
and the ZeroMQ IPC module.

It includes tests for the receiver and sender, which will
test that the pluggable mechanism works as expected and also
test sending/receiving signals using the zeromq module.

Signed-off-by: Luis Araujo <luis.araujo@collabora.co.uk>